+++ R/utils.R
@@ -10,10 +10,10 @@
 #' @return
 #' A logical vector of the same length as \code{x}.
 #'
-#' @export
-#'
 #' @examples
+#' \dontrun{
 #' is_number(c(0, 1, 1.5))
+#' }
 #'
 #' @keywords
 #' internal utils
@@ -38,11 +38,11 @@ is_number <- function(x) {
 #' Either the value of \code{expr} or in case of a failure an object of class
 #' \code{fail}, which contains the error message.
 #'
-#' @export
-#'
 #' @examples
+#' \dontrun{
 #' try_silent(1 + 1)
 #' try_silent(1 + "1")
+#' }
 #'
 #' @keywords
 #' internal utils
@@ -65,7 +65,7 @@ try_silent <- function(expr) {
 #' This function is a wrapper for \code{\link[R.utils]{withTimeout}}.
 #'
 #' @param expr
-#' An R expression to evaluate.
+#' An R expression to be evaluated.
 #' @param secs
 #' The number of seconds.
 #'
@@ -73,19 +73,19 @@ try_silent <- function(expr) {
 #' Either the value of \code{expr} or \code{NULL} if the evaluation time
 #' exceeded \code{secs} seconds.
 #'
-#' @export
-#'
 #' @examples
+#' \dontrun{
 #' foo <- function(x) { for(i in 1:10) Sys.sleep(x/10); return(x) }
 #' timed(foo(0.5), 1)
 #' timed(foo(1.5), 1)
+#' }
 #'
 #' @keywords
 #' internal utils
 
 timed <- function(expr, secs) {
   if (!(length(secs) == 1 && is_number(secs))) {
-    stop("'secs' must be a number.")
+    stop("'secs' must be a number.", call. = FALSE)
   }
   setTimeLimit(cpu = secs, elapsed = secs, transient = TRUE)
   on.exit({
@@ -99,45 +99,10 @@ timed <- function(expr, secs) {
       if (grepl("reached elapsed time limit|reached CPU time limit", e$message)) {
         return(NULL)
       } else {
-        stop(e)
+        stop(e, call. = FALSE)
       }
     }
   )
-}
-
-#' Interruption of silently tried evaluations
-#'
-#' @description
-#' This function tries to evaluate \code{expr} and interrupts the try after
-#' \code{secs} seconds.
-#'
-#' @details
-#' This function is a wrapper for \code{\link{try_silent}} in combination with
-#' \code{\link{timed}}.
-#'
-#' @param expr
-#' An R expression to try.
-#' @param secs
-#' The number of seconds.
-#'
-#' @return
-#' Either the value of \code{expr}, or (in case of a failure) an object of class
-#' \code{fail} (which contains the error message), or \code{NULL} (if the
-#' evaluation time exceeded \code{secs} seconds).
-#'
-#' @export
-#'
-#' @examples
-#' foo <- function(x, y) { Sys.sleep(x); return(x + y) }
-#' try_silent_timed(foo(0.5, 1), 1)
-#' try_silent_timed(foo(0.5, "1"), 1)
-#' try_silent_timed(foo(1.5, 1), 1)
-#'
-#' @keywords
-#' internal utils
-
-try_silent_timed <- function(expr, secs) {
-  try_silent(expr = timed(expr = expr, secs = secs))
 }
 
 #' Measure computation time
@@ -157,12 +122,12 @@ try_silent_timed <- function(expr, secs) {
 #' A list of the two elements \code{"res"} (the results of the \code{do.call}
 #' call) and \code{"time"} (the computation time).
 #'
-#' @export
-#'
 #' @examples
+#' \dontrun{
 #' what <- function(s) { Sys.sleep(s); return(s) }
 #' args <- list(s = 1)
 #' do.call_timed(what = what, args = args)
+#' }
 #'
 #' @keywords
 #' internal utils