+++ tests/testthat/test-functions.R
@@ -1,3 +1,0 @@
-test_that("Ackley function call works", {
-  expect_equal(f_ackley(c(0, 0)), 0)
-})