+++ NEWS.md
@@ -1,3 +1,15 @@
+# optimizeR 0.2.0
+
+* Some functions were renamed:
+
+  * `optimizeR()` -> `apply_optimizer()`
+  * `set_optimizer_nlm()` -> `optimizer_nlm()`
+  * `set_optimizer_optim()` -> `optimizer_optim()`
+
+* Functions `is_number()`, `try_silent()`, `timed()`, and `do.call_timed()` are no longer exported.
+
+* Function `try_silent_timed()` was removed.
+
 # optimizeR 0.1.0
 
 * Initial version.